+++ includes/class-tmsm-aquatonic-course-booking.php
@@ -249,7 +249,7 @@ class Tmsm_Aquatonic_Course_Booking {
 				$this->loader->add_filter( 'gform_pre_render_' . $form_cancel_id, $plugin_public, 'gform_pre_render_cancel', 20, 1 );
 				$this->loader->add_action( 'gform_after_submission_' . $form_cancel_id, $plugin_public, 'gform_after_submission_cancel', 20, 2 );
 			}
-			$this->loader->add_filter( 'gform_field_validation', $plugin_public, 'gform_field_validation_birthdate', 30, 4 );
+			$this->loader->add_filter( 'gform_field_validation', $plugin_public, 'gform_field_validation_add', 30, 4 );
 			$this->loader->add_filter( 'gform_pre_render_' . $form_cancel_id, $plugin_public, 'gform_pre_render_cancel', 20, 1 );
 		}
 		$this->loader->add_filter( 'gform_replace_merge_tags', $plugin_public, 'gform_replace_merge_tags_booking', 20, 7 );